rename server.js → app.js
